(#301) Windows now builds with warning level 4 (previously level 3) and treats all compiler warnings as errors; fix all level 4 warnings. Fix LTP RedPartReceptionCallback where the isEndOfBlock parameter now reports if the EOB came from the red data (correct behavior) instead of whether the last segment received from any part of the red data was the EOB (previous incorrect behavior).  Fix LTP GreenPartSegmentArrivalCallback where the offsetStartOfBlock parameter incorrectly had the length of the block added to it.

+++ common/config/include/InductsConfig.h
@@ -73,7 +73,7 @@ struct induct_element_config_t {
     uint64_t remoteNodeId;
 
     //specific to stcp and tcpcl
-    uint32_t keepAliveIntervalSeconds;
+    uint16_t keepAliveIntervalSeconds;
 
     //specific to tcpcl version 3 (servers)
     uint64_t tcpclV3MyMaxTxSegmentSizeBytes;
